Split the modal out into its own component.
