+++ web/app/containers/login/container.js
@@ -6,6 +6,7 @@ import {fetchLoginData, navigateToSection} from '../../integration-manager/login
 
 import SignInPanel from './partials/signin-panel'
 import RegisterPanel from './partials/register-panel'
+import RememberMeModal from './partials/remember-me-modal'
 
 import {Tabs, TabsPanel} from 'progressive-web-sdk/dist/components/tabs'
 
@@ -58,18 +59,21 @@ class Login extends React.Component {
                             <RegisterPanel />
                         </TabsPanel>
                     </Tabs>
+                    <RememberMeModal />
                 </div>
             )
         } else if (routeName === SIGN_IN_SECTION) {
             return (
                 <div className="t-login">
                     <SignInPanel />
+                    <RememberMeModal />
                 </div>
             )
         } else if (routeName === REGISTER_SECTION) {
             return (
                 <div className="t-login">
                     <RegisterPanel />
+                    <RememberMeModal />
                 </div>
             )
         } else {
